GAP: Update wids 102, 103, 33, and 77
The global variable `GAP_TEST_WID_102_ROUND` is used to save the
current number times of the calling of `hdl_wid_102()`. The default
value of the `GAP_TEST_WID_102_ROUND` should be 0 when the test case
is starting.

But when testing more than one case, the initialization value of the
variable GAP_TEST_WID_102_ROUND is not 0, because the previous status
is not cleared.

Add a state `mmi_round` to the class `Gap` to save the round of the mmi
be called. When a new test case is started, the class `Gap` will be re-
initialized with a new object. Than the previous state `mmi_round` will
be cleared. The state `mmi_round` is a initial state.

Same issue also found in wid 103, 33, and 77. Apply the same changes
for the wids 102, 103, 33, and 77.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>